Rename UI workspace to workbench
