+++ frontend/src/components/workbench/canvas/udf-node.tsx
@@ -30,7 +30,7 @@ import {
 import { Separator } from "@/components/ui/separator"
 import { useToast } from "@/components/ui/use-toast"
 import { getIcon } from "@/components/icons"
-import { CustomFloatingHandle } from "@/components/workspace/canvas/custom-handle"
+import { CustomFloatingHandle } from "@/components/workbench/canvas/custom-handle"
 
 export interface UDFNodeData {
   type: string // alias for key
